feat: Edit proposer dialog
